refactor: move store and result types to scoped folders and remove from types export barrels
