--- server/src/features/projects/services/cache/project-cache-page-service/index.ts
@@ -1,1 +1,0 @@
-export * from "./project-cache-page-service";
