+++ server/src/features/projects/services/project-data-service/get-project-data-key.ts
@@ -1,7 +1,8 @@
 import { CACHE } from "@/data";
 import { HashService } from "@/utils";
-import type { ProjectFilter } from "../../types";
 import { ProjectFilterUtil } from "../../utils";
+
+import type { ProjectFilter } from "../../types";
 
 /**
  * @param filenameGenerator
