+++ server/src/features/projects/services/project-tags-data-service/project-tags-data-service.ts
@@ -1,9 +1,10 @@
 import { fail } from "@/utils";
 import { ProjectTag } from "../../errors";
-import type { ProjectTagsResult } from "../../types";
 import * as DbFetcher from "../project-tags-db-fetcher.service";
 import { fetchProjectTagsData } from "./fetch-project-tags-data";
 import * as CacheManager from "./project-tags-cache-manager";
+
+import type { ProjectTagsResult } from "../../types/result";
 
 export async function createProjectTagsDataService() {
   const cacheManager = CacheManager.createProjectsTagCacheManager();
