+++ server/src/features/projects/types/index.ts
@@ -1,9 +1,5 @@
 export * from "./legacy-project-tags-cache.interface";
 export * from "./legacy-projects-cache.interface";
-export * from "./project-data-service.type";
 export * from "./project-filter.type";
 export * as Tables from "./project-tables.type";
-export * from "./project-tag-data-service.type";
-export * from "./project-tag-store-models.type";
 export * as ViewModels from "./project-viewmodels.type";
-export * from "./project-store-models.type";
