+++ server/src/features/projects/types/result/index.ts
@@ -1,0 +1,2 @@
+export * from "./project-data-service.type";
+export * from "./project-tag-data-service.type";
