+++ server/src/features/projects/types/result/project-data-service.type.ts
@@ -1,7 +1,7 @@
 import { ResultFail, ResultSuccess } from "@/types";
-import { Project } from "../errors";
-import { ProjectStoreModels } from ".";
-import { DataSource } from "./data-source.type";
+import { Project } from "../../errors";
+import { ProjectStoreModels } from "../store";
+import { DataSource } from "../data-source.type";
 
 export type ProjectResult =
   | ResultSuccess<ProjectStoreModels.Store, DataSource>
