+++ server/src/features/projects/types/result/project-tag-data-service.type.ts
@@ -1,7 +1,7 @@
 import { ResultFail, ResultSuccess } from "@/types";
-import { ProjectTag } from "../errors";
-import { DataSource } from "./data-source.type";
-import { ProjectTagStoreModels } from "./project-tag-store-models.type";
+import { ProjectTag } from "../../errors";
+import { ProjectTagStoreModels } from "../store";
+import { DataSource } from "../data-source.type";
 
 export type ProjectTagsResult =
   | ResultSuccess<ProjectTagStoreModels.Store, DataSource>
