+++ server/src/features/projects/types/store/index.ts
@@ -1,0 +1,2 @@
+export * from "./project-store-models.type";
+export * from "./project-tag-store-models.type";
